update the test to use mocha-chai-sinon, add plugins
