+++ tests/dispatch.spec.ts
@@ -1,4 +1,5 @@
 import { CustomElement, Dispatch, DispatchEmitter, Listen } from 'custom-elements-ts';
+import { expect } from 'chai'
 
 @CustomElement({
   tag: 'btn-dispatch',
@@ -10,44 +11,45 @@ class DispatchElement extends HTMLElement {
   @Dispatch('btn.namedClick') btnClickNamed: DispatchEmitter;
 
   constructor() {
-    super();
+    super()
   }
 
   @Listen('click')
   btnHandler(){
-    this.shadowRoot.querySelector('button').innerHTML = 'Hello';
-    this.btnClick.emit({detail: 'Hello'});
-    this.btnClickNamed.emit({detail: 'Hello from named click'});
+    this.shadowRoot.querySelector('button').innerHTML = 'Hello'
+    this.btnClick.emit({detail: 'Hello'})
+    this.btnClickNamed.emit({detail: 'Hello from named click'})
   }
-
 }
 
 describe('dispatch decorators', () => {
-  let element: any;
+  let element: HTMLElement
 
   beforeEach(() => {
-    const btnElement = document.createElement('btn-dispatch');
-    element = document.body.appendChild(btnElement);
-  });
+    element = document.createElement('btn-dispatch');
+    document.body.appendChild(element);
+  })
 
   afterEach(() => {
-    document.body.innerHTML = '';
-  });
+    document.body.removeChild(element)
+  })
 
   it('should trigger a btn.click DispatchEmitter', (done) => {
     element.addEventListener('btn.click', (e) => {
-      expect(e.detail).toBe('Hello');
-      done();
-    });
-    element.click();
-  });
+      /// @ts-ignore
+      expect(e.detail).equal('Hello');
+      done()
+    })
+    element.click()
+  })
 
   it('should trigger a btn.namedClick DispatchEmitter', (done) => {
     element.addEventListener('btn.namedClick', (e) => {
-      expect(e.detail).toBe('Hello from named click');
-      done();
-    });
-    element.click();
-  });
+      /// @ts-ignore
+      expect(e.detail).equal('Hello from named click');
+      done()
+    })
+    element.click()
+  })
 
-});
+})
