+++ tests/listen.spec.ts
@@ -1,122 +1,133 @@
-import { CustomElement, Dispatch, DispatchEmitter, Listen } from 'custom-elements-ts';
+import { CustomElement, Dispatch, DispatchEmitter, Listen } from 'custom-elements-ts'
 
-@CustomElement({
-  tag: 'btn-listen-dispatch',
-  template: `<button>Test</button>`
-})
-class ButtonElement extends HTMLElement {
-
-  @Dispatch() btnClick: DispatchEmitter;
-  @Dispatch('btn.namedClick') btnClickNamed: DispatchEmitter;
-
-  constructor() {
-    super();
-  }
-
-  @Listen('click')
-  btnHandler(){
-    this.shadowRoot.querySelector('button').innerHTML = 'Hello';
-  }
-
-  @Listen('click', 'button')
-  btnInnerClick() {}
-
-}
+import * as sinon from 'sinon'
+import { expect } from 'chai'
 
 describe('listen decorator', () => {
-  let element: any;
+  let element: any
+
+  @CustomElement({
+    tag: 'btn-listen-dispatch',
+    template: `<button>Test</button>`
+  })
+  class ButtonElement extends HTMLElement {
+  
+    @Dispatch() btnClick: DispatchEmitter;
+    @Dispatch('btn.namedClick') btnClickNamed: DispatchEmitter;
+  
+    constructor() {
+      super();
+    }
+  
+    @Listen('click')
+    btnHandler(){
+      this.shadowRoot.querySelector('button').innerHTML = 'Hello';
+    }
+  
+    @Listen('click', 'button')
+    btnInnerClick() {}
+  
+  }
 
   beforeEach(() => {
-    const btnElement = document.createElement('btn-listen-dispatch');
-    element = document.body.appendChild(btnElement);
+    element = document.createElement('btn-listen-dispatch');
+    document.body.appendChild(element);
   });
 
   afterEach(() => {
-    document.body.innerHTML = '';
+    document.body.removeChild(element)
+    sinon.restore()
   });
 
   it('should call method decorated @Listen', () => {
-    const btnHandlerSpy = spyOn(element.btnHandler,'call');
-    element.click();
-    expect(btnHandlerSpy).toHaveBeenCalled();
-  });
+    const btnHandlerSpy = sinon.spy(element.btnHandler,'call')
+
+    element.click()
+    sinon.assert.called(btnHandlerSpy)
+  })
 
   it('should call method decorated @Listen on children element click', () => {
-    const btnHandlerSpy = spyOn(element.btnHandler,'call');
-    element.shadowRoot.querySelector('button').click();
-    expect(btnHandlerSpy).toHaveBeenCalled();
-  });
+    const btnHandlerSpy = sinon.spy(element.btnHandler, 'call')
+
+    element.shadowRoot.querySelector('button').click()
+    sinon.assert.called(btnHandlerSpy)
+  })
 
   it('should call method decorated @Listen on selector click', () => {
-    const btnInnerSpy = spyOn(element.btnInnerClick,'call');
-    element.shadowRoot.querySelector('button').click();
-    expect(btnInnerSpy).toHaveBeenCalled();
+    const btnInnerSpy = sinon.spy(element.btnInnerClick,'call')
+
+    element.shadowRoot.querySelector('button').click()
+    sinon.assert.called(btnInnerSpy)
   });
 
   it('should execute inner method decorated @Listen', () => {
     element.click();
-    expect(element.shadowRoot.querySelector('button').innerHTML).toEqual('Hello');
-  });
+    expect(element.shadowRoot.querySelector('button').innerHTML).equal('Hello');
+  })
 
-});
-
-@CustomElement({
-  tag: 'btn-listen-shadow-false',
-  template: `<button>Test</button>`,
-  shadow: false
 })
-class ShadowFalseButtonElement extends HTMLElement {
-
-  @Dispatch() btnClick: DispatchEmitter;
-  @Dispatch('btn.namedClick') btnClickNamed: DispatchEmitter;
-
-  constructor() {
-    super();
-  }
-
-  @Listen('click')
-  btnHandler(){
-    this.querySelector('button').innerHTML = 'Hello';
-  }
-
-  @Listen('click', 'button')
-  btnInnerClick() {}
-
-}
 
 describe('listen decorator no shadowroot', () => {
-  let element: any;
+  let element: any
+
+  @CustomElement({
+    tag: 'btn-listen-shadow-false',
+    template: `<button>Test</button>`,
+    shadow: false
+  })
+  class ShadowFalseButtonElement extends HTMLElement {
+  
+    @Dispatch() btnClick: DispatchEmitter;
+    @Dispatch('btn.namedClick') btnClickNamed: DispatchEmitter;
+  
+    constructor() {
+      super();
+    }
+  
+    @Listen('click')
+    btnHandler(){
+      this.querySelector('button').innerHTML = 'Hello';
+    }
+  
+    @Listen('click', 'button')
+    btnInnerClick() {}
+  
+  }
 
   beforeEach(() => {
-    const btnElement = document.createElement('btn-listen-shadow-false');
-    element = document.body.appendChild(btnElement);
-  });
+    element = document.createElement('btn-listen-shadow-false');
+    document.body.appendChild(element);
+  })
 
   afterEach(() => {
-    document.body.innerHTML = '';
-  });
+    document.body.removeChild(element)
+    sinon.restore()
+  })
 
   it('should call method decorated @Listen', () => {
-    const btnHandlerSpy = spyOn(element.btnHandler,'call');
-    element.click();
-    expect(btnHandlerSpy).toHaveBeenCalled();
-  });
+    const btnHandlerSpy = sinon.spy(element.btnHandler,'call')
+
+    element.click()
+    sinon.assert.called(btnHandlerSpy)
+  })
 
   it('should call method decorated @Listen on children element click', () => {
-    const btnHandlerSpy = spyOn(element.btnHandler,'call');
-    element.querySelector('button').click();
-    expect(btnHandlerSpy).toHaveBeenCalled();
-  });
+    const btnHandlerSpy = sinon.spy(element.btnHandler,'call')
+
+    element.querySelector('button').click()
+    sinon.assert.called(btnHandlerSpy)
+  })
 
   it('should call method decorated @Listen on selector click', () => {
-    const btnInnerSpy = spyOn(element.btnInnerClick,'call');
-    element.querySelector('button').click();
-    expect(btnInnerSpy).toHaveBeenCalled();
-  });
+    const btnInnerSpy = sinon.spy(element.btnInnerClick,'call')
+
+    element.querySelector('button').click()
+    sinon.assert.called(btnInnerSpy)
+  })
 
   it('should execute inner method decorated @Listen', () => {
-    element.click();
-    expect(element.querySelector('button').innerHTML).toEqual('Hello');
-  });
+    element.click()
+    expect(element.querySelector('button').innerHTML).equal('Hello');
+  })
 
-});
+})
