+++ tests/prop.spec.ts
@@ -1,5 +1,7 @@
 import { CustomElement, Prop } from 'custom-elements-ts';
 
+import { expect, assert } from 'chai'
+
 @CustomElement({
   tag: 'prop-element',
   template: '<span>my element</span>',
@@ -18,65 +20,67 @@ class PropElement extends HTMLElement {
 }
 
 describe('prop decorator', () => {
-  let myElementInstance;
+  let element: any
 
   beforeEach(() => {
-    const myElement = document.createElement('prop-element');
-    myElementInstance = document.body.appendChild(myElement);
+    element = document.createElement('prop-element');
+    document.body.appendChild(element);
   });
 
   afterEach(() => {
-    document.body.innerHTML = '';
+    document.body.removeChild(element)
   });
 
   it('should reflect as property', () => {
-    myElementInstance.setAttribute('name', 'Mario');
-    expect(myElementInstance.getAttribute('name')).toEqual('Mario');
-    myElementInstance.setAttribute('name', 'Luigi');
-    expect(myElementInstance.name).toEqual('Luigi');
+    element.setAttribute('name', 'Mario')
+    expect(element.getAttribute('name')).equal('Mario')
+
+    element.setAttribute('name', 'Luigi');
+    expect(element.name).equal('Luigi');
   });
 
   it('should reflect as attribute', () => {
-    myElementInstance.name = 'Aivan';
-    expect(myElementInstance.getAttribute('name')).toEqual('Aivan');
-    myElementInstance.name = 'Custom Elements';
-    expect(myElementInstance.getAttribute('name')).toEqual('Custom Elements');
+    element.name = 'Aivan'
+    expect(element.getAttribute('name')).equal('Aivan')
+
+    element.name = 'Custom Elements'
+    expect(element.getAttribute('name')).equal('Custom Elements')
   });
 
   it('should not reflect as attribute', () => {
-    myElementInstance.name = {};
-    expect(myElementInstance.getAttribute('name')).toBeFalsy();
+    element.name = {};
+    expect(element.getAttribute('name')).to.be.null
   });
 
   it('should properly set camelcase properties', () => {
-    myElementInstance.maxFileSize = 1;
-    expect(myElementInstance.maxFileSize).toEqual(1);
+    element.maxFileSize = 1;
+    expect(element.maxFileSize).equal(1);
   });
 
   it('should properly get camelcase properties as kebabcase attributes', () => {
-    myElementInstance.maxFileSize = 1;
-    expect(myElementInstance.getAttribute('max-file-size')).toEqual('1');
+    element.maxFileSize = 1;
+    expect(element.getAttribute('max-file-size')).equal('1');
   });
 
   it('should properly set kebabcase attributes as camelcase properties', () => {
-    myElementInstance.setAttribute('max-file-size', 2);
-    expect(myElementInstance.maxFileSize).toEqual(2);
+    element.setAttribute('max-file-size', 2);
+    expect(element.maxFileSize).equal(2);
   });
 
   it('should evaluate property', () => {
-    myElementInstance.objectProp = {name: 'Name'};
-    expect(myElementInstance.objectProp.name).toEqual('Name');
+    element.objectProp = {name: 'Name'};
+    expect(element.objectProp.name).equal('Name');
   });
 
   it('should change attribute on attribute change', () => {
-    myElementInstance.setAttribute('init', 'red');
-    expect(myElementInstance.init).toEqual('red');
-    expect(myElementInstance.getAttribute('init')).toEqual('red');
+    element.setAttribute('init', 'red');
+    expect(element.init).equal('red');
+    expect(element.getAttribute('init')).equal('red');
   });
 
   it('should change property via kebab attribute', () => {
-    myElementInstance.setAttribute('kebab-case', 'shawarma');
-    expect(myElementInstance.kebabCase).toEqual('shawarma');
+    element.setAttribute('kebab-case', 'shawarma');
+    expect(element.kebabCase).equal('shawarma');
   });
 
 });
